Refactor dependency injection with centralized default services pattern
